+++ src/DumpedPlaylog.ts
@@ -1,0 +1,7 @@
+import * as amf from "@akashic/amflow";
+import * as pl from "@akashic/playlog";
+
+export interface DumpedPlaylog {
+	tickList: pl.TickList;
+	startPoints: amf.StartPoint[];
+}