+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## Unreleased changes
+* playlogの終了時間を算出する `calculateFinishedTime()` 関数を追加
+
 ## 1.1.0
 * AMFlow の非同期関数を Promise でラップしたクラス `PromisifiedAMFlowProxy` とインターフェイス `PromisifiedAMFlow` を追加